[#12] 공통 exception 수정 및 불필요한 요소 삭제
