+++ support/src/main/java/com/bluebird/pipit/support/exception/AbstractGlobalExceptionHandler.java
@@ -1,0 +1,7 @@
+package com.bluebird.pipit.support.exception;
+
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+public class AbstractGlobalExceptionHandler {
+}
