+++ support/src/main/java/com/bluebird/pipit/support/exception/GeneralExceptionCode.java
@@ -3,7 +3,5 @@ package com.bluebird.pipit.support.exception;
 public enum GeneralExceptionCode implements ExceptionCode {
 	GENERAL_ERROR,
 	INVALID_PARAMETER,
-	EXTERNAL_SERVER_ERROR,
-	SERVICE_INVALID,
-	RULE_VIOLATION
+	INVALID_BUSINESS
 }
