+++ support/src/main/java/com/bluebird/pipit/support/exception/InvalidBusinessSupportException.java
@@ -1,0 +1,19 @@
+package com.bluebird.pipit.support.exception;
+
+public class InvalidBusinessSupportException extends AttributeSupportException {
+	private final ExceptionCode code;
+
+	public InvalidBusinessSupportException(String message, ExceptionCode code) {
+		super(message);
+		this.code = code;
+	}
+
+	public InvalidBusinessSupportException(String message) {
+		this(message, GeneralExceptionCode.INVALID_BUSINESS);
+	}
+
+	@Override
+	public ExceptionCode getCode() {
+		return this.code;
+	}
+}
